Moved hooks folder to root of src
