+++ src/hooks/useActionResetGrid.js
@@ -1,5 +1,5 @@
-import InitialGridConfig from "../../config/theme/initial_grid";
-import generateGridNode from "../../util/generate_grid_nodes";
+import InitialGridConfig from "../config/theme/initial_grid";
+import generateGridNode from "../util/generate_grid_nodes";
 
 const useActionResetGrid = (_, setGridNodes) => {
   const handleSelect = () => {
